The new entities are added.

+++ Spring-boot-app/src/main/java/entities/CarService.java
@@ -44,7 +44,7 @@ public class CarService {
     @NumberFormat
     @NotBlank(message = "Employee number is mandatory")
     @Column(name = "employeesNum")
-    private Integer employeesNum;
+    private Long employeesNum;
 
   /*  @OneToMany(mappedBy = "carService")
     private List<Work> works;
